Requestの作業時間の型をDoubleからIntegerに変更 ModelはDB構造が壊れる可能性があり、かつ今後変更する可能性があるため変更はRequestのみでMapperでDoubleに変更する。

+++ src/main/java/com/example/engineary/DTO/DiaryEntryRequest.java
@@ -20,7 +20,7 @@ public class DiaryEntryRequest {
     private String contents;
 
     /** 作業時間 */
-    private Double workedTime;
+    private Integer workedTime;
 
     /** 作業日時 */
     private String workedDate;
